todo text fieldを別ファイルに分割
